refactor: make readDescription function a method of descriptionOptions
